refactor(ocr): isolate recognition runtime boundaries
